refactor: 리플렉션으로 하드코딩 된 부분 리팩토링
